+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed handling of unicode characters and non-letter characters in PROPER function. (#811)
+- Fixed SUMPRODUCT function. (#810)
 
 ## [1.2.0] - 2021-09-23